requirements.txt
asgiref==3.6.0
Django==4.2
geographiclib==2.0
geopy==2.3.0
image==1.5.33
Pillow==9.5.0
six==1.16.0
sqlparse==0.4.3
